Added constructor and changed a variable name in CategoryModel.
Further work on CategoryModel on hold for the moment.
Added an other employee title.
Added 'skeleton code' for XmlPersistence.
Removed file type responsibility from MainViewModel to PersistenceHandler
Changed MainViewModel to use one instance of a dictionary to hold the CategoryViewModels.

TODO: figure out a way to use only the dictionary to present the information in the app